+++ user_management/ui/tests/test_views.py
@@ -19,7 +19,8 @@ class TestVerifyUserEmailView(RequestTestCase):
         view = self.view_class.as_view()
         response = view(request, token=token)
         self.assertEqual(response.status_code, 302)
-        self.assertEqual(response.url, '/accounts/login/?next=/')
+
+        self.assertEqual(response.url, '/accounts/login/')
 
         user = VerifyEmailUser.objects.get(pk=user.pk)