+++ user_management/ui/views.py
@@ -1,6 +1,5 @@
 from django.conf import settings
 from django.contrib import messages
-from django.contrib.auth.views import redirect_to_login
 from django.core.exceptions import PermissionDenied
 from django.utils.translation import ugettext_lazy as _
 from django.views import generic
@@ -18,12 +17,10 @@ class VerifyUserEmailView(VerifyAccountViewMixin, generic.RedirectView):
     object, for verification.  If everything lines up, it makes the user
     active.
 
-    As a RedirectView, this will return a HTTP 302 on success.  This defaults to
-    `/` but can be overridden by changing the `url` attribute in a subclass or setting
-    LOGIN_URL in your settings.
+    As a RedirectView, this will return a HTTP 302 to LOGIN_URL on success.
     """
     permanent = False
-    url = getattr(settings, 'LOGIN_URL', '/')
+    url = settings.LOGIN_URL
     success_message = _('Your email address was confirmed.')
     invalid_exception_class = InvalidExpiredToken
     permission_denied_class = PermissionDenied
@@ -35,4 +32,4 @@ class VerifyUserEmailView(VerifyAccountViewMixin, generic.RedirectView):
     def get(self, request, *args, **kwargs):
         self.activate_user()
         messages.success(request, self.success_message)
-        return redirect_to_login('/', login_url=self.url)
+        return super(VerifyUserEmailView, self).get(request, *args, **kwargs)